Stage2A/model_training/requirements.txt
Pillow==7.0.0
tensorboard==2.2.2
pytorch-lightning==0.8.4
opencv-python==4.2.0.34
matplotlib==3.1.3
numpy==1.18.1
torch==1.5.1+cu92
